+++ src/main/java/net/mcreator/spasm/block/DirtContainerBlock.java
@@ -14,11 +14,6 @@ import net.minecraft.world.level.block.*;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.item.context.BlockPlaceContext;
-import net.minecraft.world.item.component.TooltipDisplay;
-import net.minecraft.world.item.TooltipFlag;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.Item;
-import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.player.Inventory;
@@ -33,11 +28,9 @@ import net.minecraft.core.Direction;
 import net.minecraft.core.BlockPos;
 
 import net.mcreator.spasm.world.inventory.DirtContainerGUIMenu;
-import net.mcreator.spasm.init.SpasmModBlocks;
 import net.mcreator.spasm.block.entity.DirtContainerBlockEntity;
 
 import java.util.function.Function;
-import java.util.function.Consumer;
 
 import io.netty.buffer.Unpooled;
 
@@ -157,17 +150,5 @@ public class DirtContainerBlock extends Block implements EntityBlock {
 			return AbstractContainerMenu.getRedstoneSignalFromContainer(be);
 		else
 			return 0;
-	}
-
-	public static class Item extends BlockItem {
-		public Item(Item.Properties properties) {
-			super(SpasmModBlocks.DIRT_CONTAINER.get(), properties);
-		}
-
-		@Override
-		public void appendHoverText(ItemStack itemstack, Item.TooltipContext context, TooltipDisplay tooltipDisplay, Consumer<Component> componentConsumer, TooltipFlag flag) {
-			super.appendHoverText(itemstack, context, tooltipDisplay, componentConsumer, flag);
-			componentConsumer.accept(Component.translatable("block.spasm.dirt_container.description_0"));
-		}
 	}
 }