+++ src/main/java/net/mcreator/spasm/client/gui/DirtContainerGUIScreen.java
@@ -60,6 +60,7 @@ public class DirtContainerGUIScreen extends AbstractContainerScreen<DirtContaine
 
 	@Override
 	protected void extractLabels(GuiGraphicsExtractor guiGraphics, int mouseX, int mouseY) {
+		guiGraphics.text(this.font, Component.translatable("gui.spasm.dirt_container_gui.label_dirt_container"), 5, 4, -12829636, false);
 	}
 
 	@Override